# apps/chat-api/openapi/dist/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Self-Hosted Chat API",
        "version": "0.1.0",
        "description": "Публичный HTTP-контракт чата. Собирается из фрагментов backend-пакетов (packages/backend/*/openapi/) скриптом openapi/build.php."
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctumCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "laravel_session",
                "description": "Sanctum cookie-based SPA session."
            }
        },
        "schemas": {
            "ApiError": {
                "type": "object",
                "required": [
                    "code",
                    "message",
                    "details",
                    "trace_id"
                ],
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {
                        "type": "object",
                        "additionalProperties": true
                    },
                    "trace_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
        },
        "responses": {
            "ValidationError": {
                "description": "Ошибка валидации входных данных.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
...
        }
    },
    "paths": {
        "/auth/forgot-password": {
            "post": {
                "operationId": "forgotPassword",
                "tags": [
                    "auth"
                ],
                "summary": "Запрос ссылки восстановления пароля (работает только при заданной почте)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Ответ одинаков для существующего и несуществующего email.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "message"
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "description": "Превышен rate limit.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/reset-password": {
            "post": {
                "operationId": "resetPassword",
                "tags": [
                    "auth"
                ],
                "summary": "Установка нового пароля по токену",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "token",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 10
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Пароль обновлён."
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "description": "Превышен rate limit.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}